made it so there are some checks on the button and that it shows a confirmation or error based on if the class was actually signed up for
so basically at this point the button will only reload the page if no checkbox is selected. if a checkbox is selected then it is able to go to the signup page but it first saves the data.

there is a check to make sure the user doesn't sign up for a full class, if the sign up is successful, i store the id so that later on the signup page i can show that that class was signed up for. i do the same for the ones that would fail.

on the signup page it shows a confirmation of what classes were and weren't signed up for based on the lists

check the comments on the views.py to get a more verbose explanation